extract config and key_manager
